requirements.txt
streamlit==1.37.1
supabase
python-dateutil
pandas
